requirements.txt
confluent-kafka==2.3.0
boto3
aws-msk-iam-sasl-signer-python
opensearch-py
requests
